+++ src/Slist.hs
@@ -1,11 +1,15 @@
 {-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE CPP          #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE ViewPatterns #-}
+
+{- HLINT ignore "Use mconcat" -}
 
 {- |
-Copyright:  (c) 2019 vrom911
+Copyright:  (c) 2019-2020 Veronika Romashkina
+            (c) 2020 Kowainik
 License:    MPL-2.0
-Maintainer: Veronika Romashkina <vrom911@gmail.com>
+Maintainer: Kowainik <xrom.xkov@gmail.com>
 
 This module introduces sized list data type — 'Slist'. The data type
 has the following shape:
@@ -97,6 +101,8 @@ module Slist
        , safeLast
        , init
        , tail
+       , cons
+       , cons'
        , uncons
 
          -- * Transformations
@@ -191,6 +197,15 @@ module Slist
        , sortOn
        , insert
        , insertBy
+
+        -- * Generic functions
+       , genericLength
+       , genericTake
+       , genericDrop
+       , genericSplitAt
+       , genericAt
+       , genericUnsafeAt
+       , genericReplicate
        ) where
 
 import Control.Applicative (Alternative (empty, (<|>)), liftA2)
@@ -217,7 +232,7 @@ Size can be both finite or infinite, it is established using
 data Slist a = Slist
     { sList :: [a]
     , sSize :: Size
-    } deriving (Show, Read)
+    } deriving stock (Show, Read)
 
 {- | Equality of sized lists is checked more efficiently
 due to the fact that the check on the list sizes can be
@@ -254,7 +269,7 @@ instance Monoid (Slist a) where
       where
         -- foldr :: (a -> ([a], Size) -> ([a], Size)) -> ([a], Size) -> [Slist a] -> ([a], Size)
         f :: Slist a -> ([a], Size) -> ([a], Size)
-        f (Slist l s) (xL, !xS) = (xL ++ l, s + xS)
+        f (Slist l s) (xL, !xS) = (l ++ xL, s + xS)
     {-# INLINE mconcat #-}
 
 instance Functor Slist where
@@ -346,7 +361,7 @@ instance Foldable Slist where
 
 instance Traversable Slist where
     traverse :: (Applicative f) => (a -> f b) -> Slist a -> f (Slist b)
-    traverse f (Slist l s) = (\x -> Slist x s) <$> traverse f l
+    traverse f (Slist l s) = (`Slist` s) <$> traverse f l
     {-# INLINE traverse #-}
 
 instance L.IsList (Slist a) where
@@ -659,6 +674,52 @@ init sl@Slist{..} = case sSize of
     _        -> Slist (P.init sList) (sSize - 1)
 {-# INLINE init #-}
 
+{- | @O(1)@. 'cons' is 'Slist' analogue to ':' for lists.
+It adds the given element to the beginning of the list.
+
+The following property is preserved:
+
+@
+  'size' ('cons' x xs) == 'size' xs + 1
+@
+
+Examples:
+
+>>> cons 'a' $ one 'b'
+Slist {sList = "ab", sSize = Size 2}
+
+@
+>> __'cons' 0 $ 'infiniteSlist' [1..]__
+Slist {sList = [0..], sSize = 'Infinity'}
+@
+-}
+cons :: a -> Slist a -> Slist a
+cons x (Slist xs s) = Slist (x:xs) $ s + 1
+{-# INLINE cons #-}
+
+{- | @O(1)@. Strict version of the 'cons' function
+(in terms of the size evaluation).
+
+The following property is preserved:
+
+@
+  'size' ('cons'' x xs) == 'size' xs + 1
+@
+
+Examples:
+
+>>> cons' 'a' $ one 'b'
+Slist {sList = "ab", sSize = Size 2}
+
+@
+>> __cons' 0 $ 'infiniteSlist' [1..]__
+Slist {sList = [0..], sSize = 'Infinity'}
+@
+-}
+cons' :: a -> Slist a -> Slist a
+cons' x (Slist xs !s) = let !newSize = s + 1 in Slist (x:xs) newSize
+{-# INLINE cons' #-}
+
 {- | @O(1)@. Decomposes a slist into its head and tail.
 If the slist is empty, returns 'Nothing'.
 
@@ -818,7 +879,7 @@ Slist {sList = [Slist {sList = "abc", sSize = Size 3},Slist {sList = "bac", sSiz
 -}
 permutations :: Slist a -> Slist (Slist a)
 permutations (Slist l s) = Slist
-    { sList = P.map (\a -> Slist a s) $ L.permutations l
+    { sList = P.map (`Slist` s) $ L.permutations l
     , sSize = fact s
     }
   where
@@ -1262,7 +1323,7 @@ Use 'safeIsPrefixOf' instead.
 isPrefixOf :: Eq a => Slist a -> Slist a -> Bool
 isPrefixOf (Slist l1 s1) (Slist l2 s2)
     | s1 > s2 = False
-    | otherwise = L.isPrefixOf l1 l2
+    | otherwise = l1 `L.isPrefixOf` l2
 {-# INLINE isPrefixOf #-}
 
 {- | Similar to 'isPrefixOf', but never hangs on infinite lists
@@ -1276,7 +1337,7 @@ False
 safeIsPrefixOf :: Eq a => Slist a -> Slist a -> Bool
 safeIsPrefixOf sl1@(Slist _ s1) sl2@(Slist _ s2)
     | s1 == Infinity && s2 == Infinity = False
-    | otherwise = isPrefixOf sl1 sl2
+    | otherwise = sl1 `isPrefixOf` sl2
 {-# INLINE safeIsPrefixOf #-}
 
 {- |
@@ -1302,7 +1363,7 @@ Use 'safeIsSuffixOf' instead.
 isSuffixOf :: Eq a => Slist a -> Slist a -> Bool
 isSuffixOf (Slist l1 s1) (Slist l2 s2)
     | s1 > s2 = False
-    | otherwise = L.isSuffixOf l1 l2
+    | otherwise = l1 `L.isSuffixOf` l2
 {-# INLINE isSuffixOf #-}
 
 {- | Similar to 'isSuffixOf', but never hangs on infinite lists
@@ -1316,7 +1377,7 @@ False
 safeIsSuffixOf :: Eq a => Slist a -> Slist a -> Bool
 safeIsSuffixOf sl1 sl2@(Slist _ s2)
     | s2 == Infinity = False
-    | otherwise = isSuffixOf sl1 sl2
+    | otherwise = sl1 `isSuffixOf` sl2
 {-# INLINE safeIsSuffixOf #-}
 
 {- |
@@ -1342,7 +1403,7 @@ Use 'safeIsInfixOf' instead.
 isInfixOf :: Eq a => Slist a -> Slist a -> Bool
 isInfixOf (Slist l1 s1) (Slist l2 s2)
     | s1 > s2 = False
-    | otherwise = L.isInfixOf l1 l2
+    | otherwise = l1 `L.isInfixOf` l2
 {-# INLINE isInfixOf #-}
 
 {- | Similar to 'isInfixOf', but never hangs on infinite lists
@@ -1356,7 +1417,7 @@ False
 safeIsInfixOf :: Eq a => Slist a -> Slist a -> Bool
 safeIsInfixOf sl1@(Slist _ s1) sl2@(Slist _ s2)
     | s1 == Infinity && s2 == Infinity = False
-    | otherwise = isInfixOf sl1 sl2
+    | otherwise = sl1 `isInfixOf` sl2
 {-# INLINE safeIsInfixOf #-}
 
 {- |
@@ -1892,3 +1953,163 @@ insert = insertBy compare
 insertBy :: (a -> a -> Ordering) -> a -> Slist a -> Slist a
 insertBy f a Slist{..} = Slist (L.insertBy f a sList) (sSize + 1)
 {-# INLINE insertBy #-}
+
+----------------------------------------------------------------------------
+-- Generic fuctions
+----------------------------------------------------------------------------
+
+{- | @O(1)@.
+The 'genericLength' function is an overloaded version of 'length'.
+In particular, instead of returning an 'Int', it returns any type which is an
+instance of 'Num'.
+
+>>> genericLength $ one 42
+1
+>>> genericLength $ slist [1..3]
+3
+>>> genericLength $ infiniteSlist [1..]
+9223372036854775807
+-}
+genericLength :: Num i => Slist a -> i
+genericLength = fromIntegral . length
+{-# INLINE genericLength #-}
+
+{- | @O(i) | i < n@ and @O(1) | otherwise@.
+The 'genericTake' function is an overloaded version of 'take', which
+accepts any 'Integral' value as the number of elements to take.
+
+>>> genericTake 5 $ slist "Hello world!"
+Slist {sList = "Hello", sSize = Size 5}
+>>> genericTake 20 $ slist "small"
+Slist {sList = "small", sSize = Size 5}
+>>> genericTake 0 $ slist "none"
+Slist {sList = "", sSize = Size 0}
+>>> genericTake (-11) $ slist "hmm"
+Slist {sList = "", sSize = Size 0}
+>>> genericTake 4 $ infiniteSlist [1..]
+Slist {sList = [1,2,3,4], sSize = Size 4}
+-}
+genericTake :: Integral i => i -> Slist a -> Slist a
+genericTake (fromIntegral -> i) sl@Slist{..}
+    | Size i >= sSize = sl
+    | i <= 0 = mempty
+    | otherwise = Slist
+        { sList = L.genericTake i sList
+        , sSize = min sSize (Size i)
+        }
+{-# INLINE genericTake #-}
+
+{- | @O(i) | i < n@ and @O(1) | otherwise@.
+The 'genericDrop' function is an overloaded version of 'drop', which accepts
+any 'Integral' value as the number of elements to drop.
+
+>>> genericDrop 6 $ slist "Hello World"
+Slist {sList = "World", sSize = Size 5}
+>>> genericDrop 42 $ slist "oops!"
+Slist {sList = "", sSize = Size 0}
+>>> genericDrop 0 $ slist "Hello World!"
+Slist {sList = "Hello World!", sSize = Size 12}
+>>> genericDrop (-4) $ one 42
+Slist {sList = [42], sSize = Size 1}
+
+@
+>> __drop 5 $ 'infiniteSlist' [1..]__
+Slist {sList = [6..], sSize = 'Infinity'}
+@
+
+-}
+genericDrop :: Integral i => i -> Slist a -> Slist a
+genericDrop (fromIntegral -> i) sl@Slist{..}
+    | i <= 0 = sl
+    | Size i >= sSize = mempty
+    | otherwise = Slist
+        { sList = L.genericDrop i sList
+        , sSize = sSize - Size i
+        }
+{-# INLINE genericDrop #-}
+
+{- | @O(i) | i < n@ and @O(1) | otherwise@.
+The 'genericSplitAt' function is an overloaded version of 'splitAt', which
+accepts any 'Integral' value as the position at which to split.
+
+>>> genericSplitAt 5 $ slist "Hello World!"
+(Slist {sList = "Hello", sSize = Size 5},Slist {sList = " World!", sSize = Size 7})
+>>> genericSplitAt 0 $ slist "abc"
+(Slist {sList = "", sSize = Size 0},Slist {sList = "abc", sSize = Size 3})
+>>> genericSplitAt 4 $ slist "abc"
+(Slist {sList = "abc", sSize = Size 3},Slist {sList = "", sSize = Size 0})
+>>> genericSplitAt (-42) $ slist "??"
+(Slist {sList = "", sSize = Size 0},Slist {sList = "??", sSize = Size 2})
+
+@
+>> __genericSplitAt 2 $ 'infiniteSlist' [1..]__
+(Slist {sList = [1,2], sSize = 'Size' 2}, Slist {sList = [3..], sSize = 'Infinity'})
+@
+
+-}
+genericSplitAt :: Integral i => i -> Slist a -> (Slist a, Slist a)
+genericSplitAt (fromIntegral -> i) sl@Slist{..}
+    | i <= 0 = (mempty, sl)
+    | Size i >= sSize = (sl, mempty)
+    | otherwise =
+      let (l1, l2) = L.genericSplitAt i sList
+          s2 = sSize - Size i
+      in (Slist l1 $ Size i, Slist l2 s2)
+{-# INLINE genericSplitAt #-}
+
+{- | @O(i) | i < n@ and @O(1) | otherwise@.
+The 'genericAt' function is an overloaded version of 'at', which
+accepts any 'Integral' value as the position. If the element on the given
+position does not exist it will return 'Nothing'.
+
+>>> let sl = slist [1..10]
+>>> genericAt 0 sl
+Just 1
+>>> genericAt (-1) sl
+Nothing
+>>> genericAt 11 sl
+Nothing
+>>> genericAt 9 sl
+Just 10
+-}
+genericAt :: Integral i => i -> Slist a -> Maybe a
+genericAt = at . fromIntegral
+{-# INLINE genericAt #-}
+
+{- | @O(min i n)@.
+The 'genericUnsafeAt' function is an overloaded version of 'unsafeAt', which
+accepts any 'Integral' value as the position. If the element on the given
+position does not exist it throws the exception at run-time.
+
+>>> let sl = slist [1..10]
+>>> genericUnsafeAt 0 sl
+1
+>>> genericUnsafeAt (-1) sl
+*** Exception: Slist.genericUnsafeAt: negative argument
+>>> genericUnsafeAt 11 sl
+*** Exception: Slist.genericUnsafeAt: index too large
+>>> genericUnsafeAt 9 sl
+10
+-}
+genericUnsafeAt :: Integral i => i -> Slist a -> a
+genericUnsafeAt i _ | i < 0 = errorWithoutStackTrace "Slist.genericUnsafeAt: negative argument"
+genericUnsafeAt i (Slist l Infinity) = L.genericIndex l i
+genericUnsafeAt i (Slist l (Size n))
+    | i >= fromIntegral n = errorWithoutStackTrace "Slist.genericUnsafeAt: index too large"
+    | otherwise = L.genericIndex l i
+{-# INLINE genericUnsafeAt #-}
+
+{- | @O(n)@.
+The 'genericReplicate' function is an overloaded version of 'replicate',
+which accepts any 'Integral' value as the number of repetitions to make.
+
+>>> genericReplicate 3 'o'
+Slist {sList = "ooo", sSize = Size 3}
+>>> genericReplicate (-11) "hmm"
+Slist {sList = [], sSize = Size 0}
+-}
+genericReplicate :: Integral i => i -> a -> Slist a
+genericReplicate n x
+    | n <= 0 = mempty
+    | otherwise = Slist (L.genericReplicate n x) $ Size (fromIntegral n)
+{-# INLINE genericReplicate #-}